+++ scripts/winesapos-tests.sh
@@ -1213,6 +1213,15 @@ for i in winesapos-setup.desktop winesapos-upgrade.desktop;
 done
 echo "Checking that the ${WINESAPOS_USER_NAME} user name has been set in desktop shortcuts for the setup and upgrade done."
 
+echo "Checking that the proprietary Broadcom Wi-Fi drivers are available for offline use..."
+ls -1 ${WINESAPOS_INSTALL_DIR}/var/lib/winesapos/ | grep -q broadcom-wl-dkms
+if [ $? -eq 0 ]; then
+    echo PASS
+else
+    winesapos_test_failure
+fi
+echo "Checking that the proprietary Broadcom Wi-Fi drivers are available for offline use complete."
+
 echo "Tests end time: $(date)"
 
 if (( ${failed_tests} == 0 )); then